mornings/rejoice: clarify synchronous rotation
I'm referring to the sort of planets where one side always faces the
star, like how the same side of our moon always faces Earth. The part of
such a planet right under the star would probably be too hot for (most)
life, and the eternally-dark side too cold, so the beings there probably
mostly live in the part where the sun is always low in the sky, like a
frozen sunrise/sunset, or where the sun is just below the horizon.
They'd find the very idea of the sun _moving_ pretty disturbing, maybe
even sacrilegious!

This adds slightly more detail to the poem to clarify the concept for
people unfamiliar with it.