suggestion: 'brightness here'
Before I reworked this poem to make the first line into the title and
break up the last line, [@hezzerdavis](https://twitter.com/hezzerdavis)
suggested the following modification to the last line:

> The brightness up here: our surface-dweller secret.

I removed the "up", since we're contrasting the surface not just with
the ocean depths, but also with the empty stretches of space.

Not sure which version I like more; creating this to look at them
side-by-side and sit on it a bit.