add Prayer for Productivity poem
I've been enjoying the feeling of productivity lately, but I've been
doubting that I'm actually working on anything worthwhile. Being
productive on the wrong thing is worse than being unproductive. It's
anti-productive, no matter what it feels like.

If you work on someone else's dream, you don't need to worry about that.
You can trust that your boss or CEO knows that the mission is noble, and
enjoy your toil without worry. You can be fulfilled, by the standards of
our workist culture, without wasting time thinking about if your toil
amounts to anything worthwhile.

So this is a poem that's a little bit about that.